reorganised depparser code to make space for hls See #69
